+++ app/Models/Txt.php
@@ -18,13 +18,15 @@ class Txt extends Model
 
         // Excluir o arquivo ou a pasta ao deletar o registro
         static::deleting(function ($txt) {
-            if ($txt->extension === '.zip') {
-                $extractPath = $txt->file_path.'/';
-                if (Storage::disk('public')->exists($extractPath))
-                    Storage::disk('public')->deleteDirectory($extractPath);
-            } elseif ($txt->extension === '.txt') {
-                if (Storage::disk('public')->exists($txt->file_path))
-                    Storage::disk('public')->delete($txt->file_path);
+            if($txt->isForceDeleting()){
+                if ($txt->extension === '.zip') {
+                    $extractPath = $txt->file_path.'/';
+                    if (Storage::disk('public')->exists($extractPath))
+                        Storage::disk('public')->deleteDirectory($extractPath);
+                } elseif ($txt->extension === '.txt') {
+                    if (Storage::disk('public')->exists($txt->file_path))
+                        Storage::disk('public')->delete($txt->file_path);
+                }
             }
         });
     }
